+++ controllers/auth-controllers.js
@@ -6,6 +6,7 @@ const { ctrlWrapper } = require("../utils");
 const { User } = require("../models/user");
 
 const { HttpError } = require("../helpers");
+const { date } = require("joi");
 
 const {SECRET_KEY} = process.env;
 
@@ -21,8 +22,10 @@ const register = async (req, res) => {
     
 
     res.status(201).json({
-        name: result.name,
+        // name: result.name,
         email: result.email,
+        subscription: result.subscription,
+        
     })
 }
 
@@ -41,20 +44,27 @@ const login = async (req, res) => {
     const payload = {
         id: user._id,
     }
-
+    
     const token = jwt.sign(payload, SECRET_KEY, {expiresIn: "23h"});
     await User.findByIdAndUpdate(user._id, {token});
+
     res.json ({
         token,
+        email,  
+        subscription: user.subscription
     })
 }
 
 const getCurrent = async(req, res) => {
-    const {name, email} = req.user;
+    const {
+        // name,
+         email, subscription} = req.user;
 
     res.json({
-        name,
+        // name,
         email,
+        subscription
+        
     })
 }
