+++ models/user.js
@@ -1,15 +1,16 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
+// const Joi = require("joi");
 
 const { handleMongooseError } = require("../utils");
 
+const subscriptionList = ["starter", "pro", "business"];
 const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 
 const userSchema = new Schema ({
-    name: {
-        type: String,
-        required: true,
-    },
+    // name: {
+    //     type: String,
+    //     required: true,
+    // },
     email: {
         type: String,
         match: emailRegexp,
@@ -21,6 +22,11 @@ const userSchema = new Schema ({
         minlength: 6,
         required: true,
     },
+    subscription: {
+        type: String,
+        enum: subscriptionList,
+        default: "starter"
+      },
     token: {
         type: String,
         default: null,
@@ -29,25 +35,29 @@ const userSchema = new Schema ({
 
 userSchema.post("save", handleMongooseError);
 
-const registerSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().pattern(emailRegexp).required(),
-    password: Joi.string().min(6).required(),
-});
+// const registerSchema = Joi.object({
+//     // name: Joi.string().required(),
+//     email: Joi.string().pattern(emailRegexp).required(),
+//     password: Joi.string().min(6).required(),
+// });
 
-const loginSchema = Joi.object({
-    email: Joi.string().pattern(emailRegexp).required(),
-    password: Joi.string().min(6).required(),
-});
+// const loginSchema = Joi.object({
+//     email: Joi.string().pattern(emailRegexp).required(),
+//     password: Joi.string().min(6).required(),
+// });
 
-const schemas = {
-    registerSchema,
-    loginSchema,
-}
+// const schemas = {
+//     registerSchema,
+//     loginSchema,
+// }
 
 const User = model("user", userSchema);
 
+// module.exports = {
+//     User,
+//     schemas,
+// }
+
 module.exports = {
     User,
-    schemas,
 }
