+++ routes/api/auth-routes.js
@@ -6,13 +6,16 @@ const {validateBody} = require("../../utils");
 
 const {authenticate} = require("../../middlewares");
 
- const {schemas} = require("../../models/user");
+//  const {schemas} = require("../../models/user");
+ const {registerSchema, loginSchema } = require("../../utils/validation/userValidationSchemas")
 
 const router = express.Router();
 
-router.post("/register", validateBody(schemas.registerSchema), ctrl.register);
+router.post("/register", validateBody(registerSchema), ctrl.register);
+// router.post("/register", validateBody(schemas.registerSchema), ctrl.register);
 
-router.post("/login", validateBody(schemas.loginSchema), ctrl.login);
+router.post("/login", validateBody(loginSchema), ctrl.login);
+// router.post("/login", validateBody(schemas.loginSchema), ctrl.login);
 
 router.get("/current", authenticate, ctrl.getCurrent)
