+++ utils/validateBody.js
@@ -4,7 +4,8 @@ const validateBody = (shema) => {
   const func = async (req, res, next) => {
     const { error } = shema.validate(req.body);
     if (error) {
-      next(HttpError(400, error.message));
+      next(new HttpError(400, error.message));
+      // return;
     }
     next();
   };
