+++ utils/validation/userValidationSchemas.js
@@ -1,0 +1,20 @@
+const Joi = require("joi");
+
+const registerSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().pattern(emailRegexp).required(),
+    password: Joi.string().min(6).required(),
+});
+
+const loginSchema = Joi.object({
+    email: Joi.string().pattern(emailRegexp).required(),
+    password: Joi.string().min(6).required(),
+});
+
+// const schemas = {
+//     registerSchema,
+//     loginSchema,
+// }
+
+module.exports = { registerSchema,
+    loginSchema }
